requirements.txt
clams-python==0.5.1
pyyaml==5.4.1

aniso8601==9.0.1
attrs==21.4.0
cached-property==1.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==7.1.2
deepdiff==5.2.3
Flask==1.1.4
Flask-RESTful==0.3.8
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lapps==0.0.2
lxml==4.9.0
MarkupSafe==2.0.1
mmif-python==0.4.6
ordered-set==4.0.2
platformdirs==2.5.2
pydantic==1.8.2
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2022.1
requests==2.28.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==1.0.1
zeep==4.1.0
zipp==3.1.0
